Rename Hitbox to Box
Since there will be several type of hitboxes (f.e. physical ones, defensive ones, attack ones...) and as many type of collisions, for the sake of clarity, the 'Hitbox' class name will be reserved for the combat subsystem. The physic subsystem will instead use 'Pushbox', and by renaming the current 'Hitbox' class to simply 'Box', it should clarify naming further down in dvlpt.
