+++ include/engine/box/rectangle.h
@@ -1,0 +1,21 @@
+#ifndef ENGINE_BOX_RECTANGLE_H
+#define ENGINE_BOX_RECTANGLE_H
+
+#include <iostream>
+#include <algorithm>
+
+#include "engine/box/box.h"
+
+class Rectangle: public Box
+{
+public:
+
+    Rectangle(int x, int y);
+    Rectangle(int x1, int y1, int x2, int y2);
+
+    Collision checkCollision(Rectangle* ); //implemented by this class
+    //TODO : HORRIBLE
+    Collision checkCollision(Box* h) {return dynamic_cast<Rectangle*>(h)->checkCollision(this); }; //let the other class handle the collision
+};
+
+#endif
