+++ include/game/entity/dev/red_square.h
@@ -4,7 +4,7 @@
 #include <SFML/Graphics/RectangleShape.hpp>
 
 #include "engine/entity.h"
-#include "engine/hitbox/rectangle.h"
+#include "engine/box/rectangle.h"
 
 class RedSquare: public Entity
 {
