--- src/engine/hitbox/hitbox.cpp
@@ -1,9 +1,0 @@
-#include "engine/hitbox/hitbox.h"
-
-void Hitbox::linkPos(std::pair<int, int>* pos){
-    this->pos = pos;
-}
-
-void Hitbox::linkEntity(Entity* ett){
-    this->owner = ett;
-}
