+++ src/engine/universe_master.cpp
@@ -5,7 +5,7 @@ void UniverseMaster::addEntity(Entity *entity)
     this->allEntity.push_back(entity);
     Physic *physic = entity->getPhysic();
     sf::Drawable *sprite = entity->getSprite();
-    Hitbox* hitbox = entity->getHitbox();
+    Box* hitbox = entity->getHitbox();
     if (physic)
     {
         this->allPhysic.push_back(physic);
@@ -52,7 +52,7 @@ void UniverseMaster::updateSprite()
     }
 }
 
-void UniverseMaster::checkCollision()
+std::list<Collision> UniverseMaster::checkCollision()
 {
     std::list<Collision> allCollision;
     Collision cur_coll;
@@ -68,11 +68,12 @@ void UniverseMaster::checkCollision()
             else{
                 std::cout << "Collision between " << (*it)->getOwner()->getName() << " and " << (*it2)->getOwner()->getName() << " ";
                 std::cout << "EttColliding" << cur_coll.ettColliding->getName() << ";ColVect" << cur_coll.collisionVector.first << "," << cur_coll.collisionVector.second << "\n";
+                allCollision.push_back(cur_coll);
             }
         }
         
     }
-    
+    return allCollision;
 }
 
 void UniverseMaster::update()
